avanzo altade pedido

+++ src/controllers/productoController.js
@@ -9,11 +9,17 @@ export async function getProductoById(req, res) {
   try {
     const { id } = req.params;
 
+    if (!id || isNaN(id)) {
+      const err = new Error("El id suministrado no es valido.");
+      err.status = 400;
+      throw err;
+    }
+
     const producto = await productoService.getProductoById(id);
 
     res.send(producto);
   } catch (err) {
-    res.status(404).send({
+    res.status(err.status).send({
       message: err.message,
     });
   }
@@ -27,19 +33,25 @@ export async function getProducto(req, res) {
   try {
     const queryArray = req.query;
 
-    if (queryArray === "")
-      throw new Error("No ha pasado ningun parametro a la consulta.");
+    if (queryArray === "") {
+      const err = new Error("No ha pasado ningun parametro a la consulta.");
+      err.status = 400;
+      throw err;
+    }
 
     const resp = await productoService.getProducto(queryArray);
 
-    if (resp.length === 0)
-      throw new Error(
+    if (resp.length === 0) {
+      const err = new Error(
         "No se encontraron productos con los parametros suministrados."
       );
+      err.status = 404;
+      throw err;
+    }
 
     res.send(resp);
   } catch (err) {
-    res.status(404).send({
+    res.status(err.status).send({
       message: err.message,
     });
   }
@@ -53,11 +65,19 @@ export async function createProducto(req, res) {
   try {
     const producto = req.body;
 
+    const isEmpty = Object.keys(producto).length !== 5;
+
+    if (isEmpty) {
+      const err = new Error("No se pasaron los parametros correctos.");
+      err.status = 400;
+      throw err;
+    }
+
     const resp = await productoService.createProducto(producto);
 
     res.send(resp);
   } catch (err) {
-    res.status(404).send({
+    res.status(err.status).send({
       message: err.message,
     });
   }
@@ -71,11 +91,17 @@ export async function deleteProducto(req, res) {
   try {
     const id = Number(req.params.id);
 
+    if (!id || isNaN(id)) {
+      const err = new Error("El id suministrado no es valido.");
+      err.status = 400;
+      throw err;
+    }
+
     const resp = await productoService.deleteProducto(id);
 
     res.send(resp);
   } catch (err) {
-    res.status(404).send({
+    res.status(err.status).send({
       message: err.message,
     });
   }
